requirements.txt
pydub==0.25.1
pymongo==4.4.1
openai==0.27.8
python-dotenv==1.0.0